backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
pydantic-settings==2.1.0
pymongo==4.6.0
motor==3.3.2
chromadb==0.4.15
langchain==0.1.0
langchain-google-genai==0.0.6
langchain-community==0.0.13
sentence-transformers==2.2.2
transformers==4.35.0
torch==2.1.0
pypdf2==3.0.1
python-docx==1.1.0
python-dotenv==1.0.0
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2